requirements.txt
pyspotify==2.0.0b3
SpotipyWebAPI
